backend2/requirements.txt
flake8
python_dotenv
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
alembic
httpx
pytest
pytest-asyncio
pytest-cov
pytest-mock
authlib
strands-agents
strands-agents[anthropic]